Changed video to enriched video class for intuitiveness. Also added some javadoc
